+++ CHANGES.md
@@ -5,6 +5,9 @@ x.x.x (unreleased)
 ------------------
 - Markers plugin: fix a bug where markers at the end of a track would cause
   incorrect click-to-seek behavior (#2208)
+- Regions plugin:
+  - Fix mouseup not firing if click & drag a region handle & release outside browser window (#2213)
+  - Added new `showTooltip` param allowing disabling region `title` tooltip (#2213)
 
 4.6.0 (04.03.2021)
 ------------------