+++ CHANGES.md
@@ -10,6 +10,7 @@ x.x.x (unreleased)
 - Regions plugin:
   - Fix mouseup not firing if click & drag a region handle & release outside browser window (#2213)
   - Added new `showTooltip` param allowing disabling region `title` tooltip (#2213)
+- Nullify `onaudioprocess` on remove to not execute in background (#2218)
 
 4.6.0 (04.03.2021)
 ------------------