+++ CHANGES.md
@@ -4,7 +4,17 @@ wavesurfer.js changelog
 x.x.x (unreleased)
 ------------------
 - Fixed `exportPCM()` to return valid json on `noWindow` and create a valid JSON blob
-  file on promise (#1896, #1954)
+  file returned with a Promise (#1896, #1954)
+- Playhead plugin: add a new plugin that allows the setting of a independent
+  "play head", or song-start position. (#2209)
+- Markers plugin: fix a bug where markers at the end of a track would cause
+  incorrect click-to-seek behavior (#2208)
+- Regions plugin:
+  - Fix mouseup not firing if click & drag a region handle & release outside the
+    browser window (#2213)
+  - Added new `showTooltip` param allowing disabling region `title` tooltip (#2213)
+- Nullify `onaudioprocess` on remove to not execute in background (#2218)
+- Add new `vertical` parameter enabling displaying waveforms vertically (#2195)
 
 4.6.0 (04.03.2021)
 ------------------